# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Contact:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.ContactInput:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  utils.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact:
    name: Rafi Kurnia Putra
  description: API for Contact
  license:
    name: MIT License
    url: https://github.com/rafikurnia/go-webapp/blob/main/LICENSE
  title: go-webapp
  version: 2.0.0
paths:
  /contacts:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Contact'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get a list of contacts
      tags:
      - Contacts
    post:
      consumes:
      - application/json
      parameters:
      - description: Each field is required, otherwise it will be set as an empty
          string.
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.ContactInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Create a contact
      tags:
      - Contacts
  /contacts/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: The ID of the contact
        format: uint
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Delete a contact with the specified ID
      tags:
      - Contacts
    get:
      consumes:
      - application/json
      parameters:
      - description: The ID of the contact
        format: uint
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Returns a contact with the specified ID
      tags:
      - Contacts
    put:
      consumes:
      - application/json
      parameters:
      - description: The ID of the contact
        format: uint
        in: path
        name: id
        required: true
        type: integer
      - description: Each field is required, otherwise it will be set as an empty
          string.
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.ContactInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Update a contact with the specified ID
      tags:
      - Contacts
swagger: "2.0"